Naive verbose flag
Output suppression (< /dev/null) is disabled if the verbose (-v) flag is
included as the 2nd argument in the commmand

Note: this implementation is naive. It checks positionally, and not by
consuming the entire input and tokenizing.

This implementation also only allows one form of the flag to be passed
in: -v (but not --verbose)